Change banner on start page to be consistent with the applications page. Update feature flag name to be more generic for reuse
