+++ app/views/provider_interface/start_page/show.html.erb
@@ -25,17 +25,16 @@
   </div>
 </div>
 
-<% if FeatureFlag.active?('getting_ready_for_next_cycle_banner') %>
+<% if FeatureFlag.active?('provider_information_banner') %>
   <div class="govuk-width-container">
-    <div class="app-banner govuk-!-margin-top-4 govuk-!-margin-bottom-0" aria-labelledby="getting-ready-for-next-cycle-message" >
+    <div class="app-banner govuk-!-margin-top-4 govuk-!-margin-bottom-0" aria-labelledby="change-to-working-days-message">
       <div class="app-banner__message">
-        <h2 class="govuk-heading-m" id="getting-ready-for-next-cycle-message">
-          <%= govuk_link_to(
-            'Guidance for the new cycle (2020 to 2021)',
-            provider_interface_guidance_for_the_new_cycle_path,
-            class: 'govuk-link--no-visited-state',
-          ) %>
+        <h2 class="govuk-heading-m" id="change-to-working-days-message">
+          20 December 2020 to 1 January 2021 do not count as working days
         </h2>
+        <p class="govuk-body">
+          The Christmas period will not be included when calculating dates for applications to be automatically rejected. This is to match the dates used by UCAS.
+        </p>
       </div>
     </div>
   </div>
